requirements.txt
selenium
selenium-stealth
python-telegram-bot
SQLAlchemy
beautifulsoup4
requests
